add ViT approach
