+++ FLOODPY_gpu_env.yml
@@ -1,0 +1,330 @@
+name: floodpy_gpu
+channels:
+  - pytorch
+  - conda-forge
+  - defaults
+dependencies:
+  - _libgcc_mutex=0.1
+  - _openmp_mutex=4.5
+  - abseil-cpp=20211102.0
+  - affine=2.4.0
+  - appdirs=1.4.4
+  - arrow-cpp=8.0.0
+  - asttokens=2.4.0
+  - attrs=23.2.0
+  - aws-c-common=0.6.8
+  - aws-c-event-stream=0.1.6
+  - aws-checksums=0.1.11
+  - aws-sdk-cpp=1.8.185
+  - backcall=0.2.0
+  - backports=1.0
+  - backports.functools_lru_cache=1.6.5
+  - binutils_impl_linux-64=2.38
+  - blas=1.0
+  - blosc=1.21.5
+  - bokeh=3.1.1
+  - boost-cpp=1.78.0
+  - branca=0.7.2
+  - brotli=1.1.0
+  - brotli-bin=1.1.0
+  - brotlipy=0.7.0
+  - bzip2=1.0.8
+  - c-ares=1.28.1
+  - ca-certificates=2024.7.4
+  - cached-property=1.5.2
+  - cached_property=1.5.2
+  - cads-api-client=1.1.0
+  - cairo=1.16.0
+  - cdsapi=0.7.0
+  - certifi=2024.7.4
+  - cffi=1.15.1
+  - cfitsio=4.2.0
+  - cftime=1.6.3
+  - charset-normalizer=2.0.4
+  - click=8.1.7
+  - click-plugins=1.1.1
+  - cligj=0.7.2
+  - cloudpickle=3.0.0
+  - colorama=0.4.6
+  - colorcet=3.1.0
+  - comm=0.1.4
+  - cryptography=41.0.2
+  - cudatoolkit=11.1.1
+  - curl=8.1.2
+  - cytoolz=0.12.3
+  - dask=2023.5.0
+  - dask-core=2023.5.0
+  - datashader=0.15.2
+  - datashape=0.5.4
+  - debugpy=1.6.7
+  - decorator=5.1.1
+  - distributed=2023.5.0
+  - einops=0.8.0
+  - entrypoints=0.4
+  - executing=1.2.0
+  - expat=2.6.2
+  - fiona=1.9.1
+  - folium=0.17.0
+  - font-ttf-dejavu-sans-mono=2.37
+  - font-ttf-inconsolata=3.000
+  - font-ttf-source-code-pro=2.038
+  - font-ttf-ubuntu=0.83
+  - fontconfig=2.14.2
+  - fonts-conda-ecosystem=1
+  - fonts-conda-forge=1
+  - freetype=2.12.1
+  - freexl=1.0.6
+  - gcc=12.1.0
+  - gcc_impl_linux-64=12.1.0
+  - gdal=3.6.2
+  - geopandas=0.13.2
+  - geopandas-base=0.13.2
+  - geos=3.11.1
+  - geotiff=1.7.1
+  - gettext=0.21.1
+  - gflags=2.2.2
+  - giflib=5.2.2
+  - glog=0.7.1
+  - grpc-cpp=1.46.3
+  - h5py=3.8.0
+  - hdf4=4.2.15
+  - hdf5=1.12.2
+  - icu=70.1
+  - idna=3.4
+  - imagecodecs-lite=2019.12.3
+  - imageio=2.34.2
+  - importlib-metadata=8.0.0
+  - importlib_metadata=8.0.0
+  - importlib_resources=6.4.0
+  - intel-openmp=2023.1.0
+  - ipykernel=6.25.2
+  - ipython=8.12.2
+  - jbig=2.1
+  - jedi=0.19.0
+  - jinja2=3.1.4
+  - joblib=1.4.2
+  - jpeg=9e
+  - json-c=0.16
+  - jupyter_client=7.3.4
+  - jupyter_core=5.3.0
+  - kealib=1.5.0
+  - kernel-headers_linux-64=2.6.32
+  - keyutils=1.6.1
+  - kiwisolver=1.4.5
+  - krb5=1.20.1
+  - lazy_loader=0.4
+  - lcms2=2.15
+  - ld_impl_linux-64=2.38
+  - lerc=4.0.0
+  - libabseil=20211102.0
+  - libaec=1.1.3
+  - libblas=3.9.0
+  - libbrotlicommon=1.1.0
+  - libbrotlidec=1.1.0
+  - libbrotlienc=1.1.0
+  - libcblas=3.9.0
+  - libcurl=8.1.2
+  - libdap4=3.20.6
+  - libdeflate=1.17
+  - libedit=3.1.20191231
+  - libev=4.33
+  - libevent=2.1.10
+  - libexpat=2.6.2
+  - libffi=3.4.4
+  - libgcc-devel_linux-64=12.1.0
+  - libgcc-ng=13.2.0
+  - libgdal=3.6.2
+  - libgfortran-ng=11.2.0
+  - libgfortran5=11.2.0
+  - libglib=2.78.1
+  - libgomp=13.2.0
+  - libiconv=1.17
+  - libkml=1.3.0
+  - liblapack=3.9.0
+  - libllvm14=14.0.6
+  - libnetcdf=4.9.1
+  - libnghttp2=1.58.0
+  - libnsl=2.0.1
+  - libpng=1.6.43
+  - libpq=15.2
+  - libprotobuf=3.20.3
+  - librttopo=1.1.0
+  - libsanitizer=12.1.0
+  - libsodium=1.0.18
+  - libspatialindex=2.0.0
+  - libspatialite=5.0.1
+  - libsqlite=3.45.3
+  - libssh2=1.11.0
+  - libstdcxx-ng=13.2.0
+  - libthrift=0.15.0
+  - libtiff=4.5.0
+  - libuuid=2.38.1
+  - libuv=1.44.2
+  - libwebp-base=1.4.0
+  - libxcb=1.13
+  - libxml2=2.10.3
+  - libzip=1.10.1
+  - libzlib=1.2.13
+  - llvmlite=0.40.1
+  - locket=1.0.0
+  - lz4=4.3.3
+  - lz4-c=1.9.3
+  - mapclassify=2.5.0
+  - markupsafe=2.1.5
+  - matplotlib-base=3.7.3
+  - matplotlib-inline=0.1.6
+  - mkl=2023.1.0
+  - mkl-service=2.4.0
+  - mkl_fft=1.3.6
+  - mkl_random=1.2.2
+  - msgpack-python=1.0.8
+  - multipledispatch=0.6.0
+  - multiurl=0.3.1
+  - munch=4.0.0
+  - munkres=1.1.4
+  - ncurses=6.4
+  - nest-asyncio=1.5.6
+  - netcdf4=1.6.3
+  - networkx=3.1
+  - ninja=1.10.2
+  - ninja-base=1.10.2
+  - nspr=4.35
+  - nss=3.98
+  - numba=0.57.1
+  - numpy-base=1.24.3
+  - openjpeg=2.5.0
+  - openssl=3.1.6
+  - orc=1.7.4
+  - packaging=23.1
+  - pandas=1.4.2
+  - param=1.13.0
+  - parso=0.8.3
+  - partd=1.4.1
+  - pcre=8.45
+  - pcre2=10.40
+  - pexpect=4.8.0
+  - pickleshare=0.7.5
+  - pip=23.2.1
+  - pixman=0.40.0
+  - platformdirs=3.10.0
+  - pooch=1.4.0
+  - poppler=23.03.0
+  - poppler-data=0.4.12
+  - postgresql=15.2
+  - proj=9.1.1
+  - prompt-toolkit=3.0.39
+  - prompt_toolkit=3.0.39
+  - pthread-stubs=0.4
+  - ptyprocess=0.7.0
+  - pure_eval=0.2.2
+  - pyarrow=8.0.0
+  - pycparser=2.21
+  - pyct=0.5.0
+  - pygments=2.16.1
+  - pyopenssl=23.2.0
+  - pyproj=3.5.0
+  - pysocks=1.7.1
+  - python=3.8.17
+  - python-dateutil=2.8.2
+  - python_abi=3.8
+  - pytorch=1.8.1
+  - pytz=2024.1
+  - pywavelets=1.4.1
+  - pyyaml=6.0.1
+  - rasterio=1.3.6
+  - re2=2022.04.01
+  - readline=8.2
+  - requests=2.31.0
+  - rioxarray=0.13.4
+  - rtree=1.3.0
+  - scikit-image=0.20.0
+  - scikit-learn=1.3.2
+  - setuptools=68.0.0
+  - shapely=2.0.1
+  - six=1.16.0
+  - snappy=1.2.0
+  - snuggs=1.4.7
+  - sortedcontainers=2.4.0
+  - sqlite=3.41.2
+  - stack_data=0.6.2
+  - sysroot_linux-64=2.12
+  - tbb=2021.8.0
+  - tblib=3.0.0
+  - threadpoolctl=3.5.0
+  - tiledb=2.13.2
+  - tk=8.6.12
+  - toolz=0.12.1
+  - tqdm=4.66.4
+  - traitlets=5.9.0
+  - typing-extensions=4.7.1
+  - typing_extensions=4.7.1
+  - tzcode=2024a
+  - tzdata=2024a
+  - unicodedata2=15.1.0
+  - uriparser=0.9.7
+  - urllib3=1.26.16
+  - utf8proc=2.6.1
+  - wcwidth=0.2.6
+  - wheel=0.38.4
+  - xarray=2023.1.0
+  - xarray-spatial=0.4.0
+  - xbatcher=0.3.0
+  - xerces-c=3.2.4
+  - xorg-kbproto=1.0.7
+  - xorg-libice=1.1.1
+  - xorg-libsm=1.2.4
+  - xorg-libx11=1.8.4
+  - xorg-libxau=1.0.11
+  - xorg-libxdmcp=1.1.3
+  - xorg-libxext=1.3.4
+  - xorg-libxrender=0.9.10
+  - xorg-renderproto=0.11.1
+  - xorg-xextproto=7.3.0
+  - xorg-xproto=7.0.31
+  - xyzservices=2024.6.0
+  - xz=5.4.2
+  - yaml=0.2.5
+  - zeromq=4.3.4
+  - zict=3.0.0
+  - zlib=1.2.13
+  - zstd=1.5.5
+  - pip:
+      - asciitree==0.3.3
+      - contourpy==1.1.0
+      - cycler==0.11.0
+      - defusedxml==0.7.1
+      - docker-pycreds==0.4.0
+      - dominate==2.6.0
+      - fasteners==0.18
+      - fonttools==4.42.1
+      - fsspec==2023.9.0
+      - gitdb==4.0.10
+      - gitpython==3.1.35
+      - imagecodecs==2023.3.16
+      - importlib-resources==6.0.1
+      - lxml==4.9.3
+      - matplotlib==3.7.2
+      - numcodecs==0.11.0
+      - numpy==1.24.4
+      - pathtools==0.1.2
+      - pillow==8.0.1
+      - promise==2.3
+      - protobuf==3.20.3
+      - psutil==5.9.5
+      - pyparsing==3.0.9
+      - pyzmq==25.1.1
+      - scipy==1.9.3
+      - sentry-sdk==1.30.0
+      - setproctitle==1.3.2
+      - shortuuid==1.0.11
+      - smmap==5.0.0
+      - tifffile==2023.7.10
+      - torchfile==0.1.0
+      - torchsummary==1.5.1
+      - torchvision==0.9.1
+      - tornado==6.3.3
+      - visdom==0.1.8
+      - wandb==0.12.18
+      - websocket-client==1.6.2
+      - zarr==2.16.1
+      - zipp==3.16.2
