+++ floodpy/Preprocessing_S1_data/DEM_funcs.py
@@ -1,9 +1,11 @@
 import xarray as xr
 from xrspatial import slope
+import os
 import rioxarray  # activate the rio accessor
 
-def calc_slope_mask(Floodpy_app):
+def calc_slope_mask(Floodpy_app, overwrite=True):
 
+    if os.path.exists(Floodpy_app.DEM_slope_filename): os.remove(Floodpy_app.DEM_slope_filename)
     DEM_xarray = xr.open_dataset(Floodpy_app.DEM_filename, decode_coords='all')
     DEM_utm_xarray = DEM_xarray.rio.reproject(DEM_xarray.rio.estimate_utm_crs())
     DEM_utm_xarray = DEM_utm_xarray.coarsen(x=6, boundary='pad').max().coarsen(y=6,boundary='pad').max()
